Adds `bItemCast=TRUE` flag to Hold weapatt spell cast (#945)
This further avoids traditional spell cast logic by flagging this
weapatt's spell cast as an item cast via `bItemCast`. Doing so avoids a
few bits of traditional spell casting code in Spell's `CastSpell()`.

In Discord, I had suggested that without this flag the player might be
improving off a weapon attribute. However, an instance of `self` (the
weapon attribute) is being passed to `CastSpell()`, standing in as the
caster, and causes `ImproveAbility()` to return immediately.