Seduce: Verify conditions before continuing enchantment (#929)
This PR fixes #461 by verifying that the target monster is still alive
during each iteration of Seduce's mana drain timer. Currently, if an
affected mob is killed, `EndEnchantmentTimer` doesn't recognize this and
the mana drain continues until the caster is out of mana. Additionally,
it performs new checks to ensure the master (player) is present.

**About Seduce**: When cast on a monster, seduce will cause the monster
to fight on your behalf for as long as your mana allows. If your mana
runs out, the monster turns on you. Seduce can be cast on multiple
monsters. If a monster is killed, the mana drain (from that monster)
should stop.