Fixes check issues in Bourgeois Castle (`castle2b`) (#952)
This fixes behind-the-scenes unclosed sector and cross-reference issues.
For example, several linedefs had sidedefs that needed to be nullified
by replacing their generated sidedef value with -1. No visual changes.

Note: The keep guards atop each tower in this room flicker because of
the skybox. AFAIK, there's nothing we can do about that short of
removing the windows or moving these keep guards to the ground.

![Screenshot 2024-09-14
084201](https://github.com/user-attachments/assets/5c0c0505-4148-4dba-ab54-c1d629203777)
![Screenshot 2024-09-14
084212](https://github.com/user-attachments/assets/6d8b0f12-41a0-4669-b47a-499ff7d53194)