Fixes territory game recreate by skipping post-processing on flag claims (#947)
This PR modifies the `SetFaction` function to only change the flag's
faction without performing post-processing steps. Specifically, the flag
claim process and other related actions will no longer occur when only
the faction is updated. This is needed to properly recreate the
territory game and specifically, the full version of the territory game.

## Testing
* I manually converted flags to another faction when testing. Using
`send o <TerritoryGame> GetFlagCounts` to verify the counts were the
expected values.
* Tested existing functionality to ensure no side effects or unintended
behaviors were introduced when recreating the territory game and when
performing a system `RecreateAll`.

## Notes:
* This has been attempted and discussed before in #450 
* This attempts to address the issue with the least change possible, so
it doesn't address some [desires in the original
issue](https://github.com/Meridian59/Meridian59/issues/29#issuecomment-1478850125).
I can address those in this or we can leave that for a follow-up.

Fixes #29 (at least what I understand to be the remaining issue)