Remove ranged weapons' impact on berserker ring durability  (#956)
This PR is an alternative to #658. Currently, when ranged weapons are
used in tandem with a berserker ring, the ring's durability will
decreases even though it's damage buff does not apply to ranged weapons.
This will remove the invisible effect of ranged weapons on the ring's
durability. It _also_ replaces some magic numbers. I can revert if
preferred.