Allows for player defiling of good monster corpses
This commit makes it possible to Defile the corpses of good monsters, in addition to good player corpses.

It borrows a player-based flag in `body` to make this possible and sets this `pbWas_good_player` flag to true when generating the corpse in `monster`. Logic was removed from `GetObjectFlags()` to make this possible.