Expires faction service quests when resigned from faction (#918)
This PR addresses an issue brought up in #465. When players are on a
service quest, but lose faction, the service quest remains active. At
best, the player will receive a late gmail about being booted from the
faction. At worst, they'll be booted from their new faction, if they
joined one in the meantime. This happens because the quest penalty calls
`ResignFaction`. This PR seeks to solve this issue by immediately
expiring the related service quest when a player is booted from a
faction.

## Testing
Due to the complexity of the quest system, this can be challenging to
test. I dialed down the quest engines `piQuestTimerDelay`,
`piQuestDeadlineTimerDelay`, and `piQuestHistoryRecentTime` to speed up
triggering the loyalty quest in quick succession, as I was jumping from
one faction to another to ensure each factions quest stats were
accurate.

### General testing steps
1. Join a faction
2. Trigger a loyalty quest
3. Immediately get yourself kicked out (e.g. attack a faction solder)
4. Confirm you've lost faction

When testing, it's most important to verify that the quest is properly
expired. Faction quest slots are limited to the number of faction
members, so it's important to check that the maximum number of available
quests in the quest template adjust with membership (+1 when joining, -1
when resigning). Also, expiring this quest in this new way shouldn't
simply delete it from history. This PR expires it quietly, allowing it
to do all the things it would normally do without triggering the gmail.
I've added a `report` flag to manage this.

I've extracted some of this functionality into new methods in
`QuestEngine` and `QuestNode` thinking they might be helpful in the
future. I looked up and down the codebase and could not find anything
that handled this currently.

### Things to consider
- The faction service quest template ids are 8 (Princess), 9 (Duke), and
69 (Rebel).

### Thoughts
- Ideally, _all_ related faction quests would be failed if booted from a
faction, but that felt like opening up a can of worms. As of today, only
the 3 service quests have a faction loss penalty and cause this issue.
- The same issue was dealt with very differently in OpenMeridian
[here](https://github.com/OpenMeridian/Meridian59/pull/479).

This closes #465