Quality of Life: Auto-delete depleted wands and Froz (#983)
This PR serves two purposes - to remove the clutter that inevitably
clogs inventories when using wands, or Jewels of Froz; and bringing
consistency to the way that consumable combat items are disposed of.

Both wands and Froz remove delete themselves after depletion, like
Berserker Rings. The "shatters" message for wands already makes perfect
sense, but the Froz depletion message was updated to be more accurate
and lore friendly.