Adds timer duration to DM created weapon spell attributes
This commit addresses an issue with DM-added weapon attributes of the spell attack type. Currently, when these attributes are added via `dm create itematt holy`, for example, a timer is not provided. It's needed by a later check and it's absence causes the errors seen in #1007.

These weapon attributes are randomly generated with a duration of 1-24 hours. For the DM version, I chose 24-hours as anything created by a god ought to last a while.