Adds 24-hour duration to DM-added weapon spell attributes (#1014)
This PR addresses an issue with DM-added weapon attributes of the spell
attack type. Currently, when these attributes are added via `dm create
itematt holy`, for example, a timer is not provided. It's needed by a
later check and its absence causes the errors seen in #1007.

These weapon attributes are randomly generated with a duration of 1-24
hours. For the DM version, I chose 24-hours as anything created by a god
ought to last a while.

I also snuck in a small change to the spell type descriptions that get
appended for consistency with all other text formatting.

### Testing
* `dm get item mace`
* Equip the mace
* `dm create itematt holy`
* Examine the mace to verify it's enchanted with a holy attack
* Verify the error logs are clear of issues


This fixes #1007