Fix for Portuguese client account creation (#1023)
This addresses a critical issue with the Portuguese translations and
tackles some housekeeping:

1. Adds Portuguese translations for the signup dialog, which do not
currently exist. This includes incorrect entries for account creation
endpoints that are currently causing account creation to render a false
positive.
2. Replaces Portuguese string table numerical ids with their constant
equivalent.
3. Replaces references to non-existent help pages with the Guides
section of the website.


![image](https://github.com/user-attachments/assets/68395086-e9b0-43ea-a777-f7b8533bed2c)


While investigating, I also noticed that English is not the language
fallback. Entries not duplicated across languages will be missing. E.g.,
the rarity translations are only present in the English client. I'm
planning to submit separate PRs to address these issues.