Upgrade existing statues (#1022)
After we added the ability to show overlays on statues:
https://github.com/Meridian59/Meridian59/pull/927

We found out that any existing statues that had overlays that were
created before this upgrade would have a mismatch between the number of
calculated overlays and the overlays sent to the client. We create a one
time upgrade to solve this, fixing any existing statues that need to be
recalibrated.

Testing:

Created a server pre-upgrade and reset the statues and saved the game
state. Upgraded post https://github.com/Meridian59/Meridian59/pull/927 -
got the bug and then ran this fix, saw the one time upgrade applied to
all statues and the bug was gone.