Fixes mail subject handling tied to recipient count (#1029)
This fixes a critical issue where mail subject handling was incorrectly
tied to the number of recipients.

Previously, the subject line could be written in the wrong position. If
there were more than three recipients, the client would crash.

To resolve this, I adjusted the logic by tracking the subject ID
separately.

This supports blank subjects, which appeared to be a problem when this
bug surfaced.