Mirror background textures on the x-axis (#1040)
Adjust the texture coordinates (s and t) of the vertices of a background
overlay to mirror the texture on the x-axis.
Since [rotating the
backgrounds](https://github.com/Meridian59/Meridian59/pull/1016) to
match the origin image orientations, the background objects were no
longer clickable. To prevent this, we revert the change and instead we
swap the s coordinates of the vertices. The background objects are once
again clickable.

before:

![image](https://github.com/user-attachments/assets/3e6ec456-6700-4e24-b2ca-defc50f3329d)

after:

![image](https://github.com/user-attachments/assets/16dcbc71-b5cf-4bcc-88d1-a7cafc75e0c5)

Fixes: #1039