Denies Shadow Rift in `NO_COMBAT` rooms (#980)
This commit adds `NO_COMBAT` rooms to the list of rooms excluded from
shadow rift portals. Previously, only the `HISTORIC` rooms were
excluded. This commit also moves the logic checking this to
`CanPayCosts()`, so that the user gets immediate feedback that it cannot
be cast.

`NO_COMBAT` rooms include:
- Inns
- Taverns
- Adventurer Halls
- Raza
- Many of the NPC/town rooms (blacksmiths, the vault, banks, etc)
- The Clearing of the Trading Post, Outside the Mad Scientist's Hut, The
Remote Trading Post of Wulfgang zax'Ak, and a few others
- Rented Rooms