Updates mollusk creature to include initial underground state (#903)
This PR enhances mollusk creatures in a couple of ways:
1. It updates the code and bgf to include a new "underground" state from
which mollusks emerge once engaged. This state was in a separate
`molrock.bgf` and it looks like some support might have been written
into `mollusk.kod`, was incomplete.
2. I extracted the bitmaps from both existing bgfs, cleaned up alpha
channel color bleed, then repackaged them in a new mollusk.bgf using
`makebgf` and an updated `bbg`.

The most challenging aspect of this was disallowing the monster from
doing anything until the popup animation completed. Some of my earlier
solutions results in out of sync animation frames. I settled on a timer
that paralyzes the monster during the animation.

#### mollusk.bgf source assets

[mollusk_assets.zip](https://github.com/user-attachments/files/16921998/mollusk_assets.zip)


## Testing
This new behavior should be tested using various triggers, including up
close melee and ranged attacks. In all cases, the mollusk should remain
in position, emerge, then start chasing and attacking the player.

### Commands
`dm monster mollusk creature` to generate a Mollusk

### The new behavior when attacked up close and from afar


https://github.com/user-attachments/assets/c8835bf7-0c46-4ba0-bff4-82168bebea11


https://github.com/user-attachments/assets/e2747d02-b765-494c-abe8-fc83cbf7a0b0


**Misc:**
Something that bothers me about the artwork is how the mollusk appears
to hover through many of its animations. This is because its club arm
extends beyond its feet. It seems, the right fix for that is updating
the artwork.

Closes #429