feature: Adds directional placement to rented room objects (#1064)
This PR updates object placement in rented rooms to align the object's
orientation with the player's facing direction. This change primarily
affects multi-sided objects, such as the personal chest. The decorator's
dialog has been updated to reflect this.

The decorator says this when placing an object is requested, **bolded**
text is newly added:
> Please stand where you'd like the object, **look in the direction it
should face,** and say 'here'."

Fixes #646