No eats under second wind (#1067)
Currently players are able to consume foods with an active second wind
enchantment. This fills up their stomach capacity and removes the food
item from their inventory but does nothing to help the player.

This update would prevent players from eating any foods - notifying them
that they are still under the surge of adrenaline from the second wind.

To keep gameplay impacts as small as possible an exception is made for
chocolate mints - these will be consumed and will increase the player's
mana but will not increase the player's vigor. In addition, potions and
flasks are eaten according to KOD but they are not food items and will
be consumed and have the normal effect.

This is originally submitted as a draft PR in the hopes of finding some
confirmation that the IsItemType() message is used correctly. This
message lives in Item.kod already and returns:

      return (viItem_type & type);
      
Where each item can have multiple types so the bitwise and is a bit
strange here to me. In testing this current set up works. I've also
considered adding a new IsFood() message in a few spots (in item it
returns false, in Food.kod we return true, etc).