fix: Allow guild allies to exit allied guild halls (#1071)
This PR fixes #703, addressing an issue where guild allies can enter an
allied guild hall but are unable exit. The root cause lies in the
inconsistent logic between entry and exit handling. Most guild halls use
`GuildHall::SomethingTryGo` for entry, but rely on their own overridden
`SomethingTryGo` for exiting. The exit logic, which uses
`ReqLegalEntry`, did not account for allies.

To resolve this, `ReqLegalEntry` has been updated to ensure that anyone
permitted to use the door to enter from the outside can also use it to
exit from the inside.

~**Updated:** For consideration, the latest commit refactors this setup
for clarity. It limits `ReqLegalEntry` to focus solely on legal entry
per the hotplate system, and updates exit handling to use `CanEnter`.~