fix: Revises the placement of braziers in Joguer's (`barapoth`) (#1110)
This PR adds three braziers to Joguer's—two that were clearly meant to
be there but were missing, and a third that was commented out. The room
has two flickering light sectors that weren’t properly matched with
light sources, and the third brazier adds a nice touch behind the
counter.


![image](https://github.com/user-attachments/assets/6ff7b933-8038-4f29-841c-7b357b18417b)


![image](https://github.com/user-attachments/assets/5ea1126b-fb23-4b19-aafc-f49da4f54af5)


![image](https://github.com/user-attachments/assets/164d6322-5c7b-4723-989d-e28e0581c779)