fix: Allow for ambient lighting in Icky, increase sector light levels (#1108)
This PR updates all of Icky's sectors to be influenced by ambient
lighting and raises their light levels from 54 to 150. This brings this
cave in line with the lighting in the adjacent orc cave (through the
hidden entrance).

I will be submitting a related PR that lowers that base light level of
Icky from LIGHT_DARK to LIGHT_MIN. That change will make it consistent
with the orc caves beyond where these light levels are used: sewers and
caves.

Fixes #1101