fix: Improvements to map annotations and tooltips (#1170)
This PR introduces a couple of improvements to map annotations behavior:

1. **Annotations and tooltips are now fully disabled if either "Show
dynamic minimap" or "Map annotations" is turned off.**
One could argue there's some benefit to having them display, but I think
it's more confusing to have tooltips appear for annotations the player
can't see.

2. **Tooltips are now properly cleared when many dialogs are
displayed.**
I added a wrapper around `DialogBoxParam` that calls an updated
`TooltipReset()` beforehand. This helps clear lingering tooltips before
most dialogs appear. I skipped a few dialogs that likely don’t interact
with tooltips.

I recorded this video to show a variety of conditions under these
changes:
- Players cannot create or edit map annotations, or see their tooltips,
if either the minimap is disabled
- Players cannot create or edit map annotations, or see their tooltips,
if either map annotations are disabled
- An annotation's tooltip will no longer remain under the mouse cursor
if an annotation is edited, or if the Look dialog is brought up. Not
shown are the other scenarios:
  - When the who dialog is opened
  - When a buy dialog is opened
  - When a newsglobe is opened
  - When the mail reader is opened
  - When an offer dialog is opened
  - When the color selector is opened


https://github.com/user-attachments/assets/a43c7be9-7ef2-4b91-9414-c4fd3da63402