Move intro DLL into main client.
If we ever switched to 64-bit compilation, we'd have a problem,
because the 64 bit client would initially try to load the 32 bit
intro.dll, which would fail.  So we move the intro DLL code into the
main client.  The separate intro DLL wasn't buying us anything anyway.

The other modules aren't a problem, because they can be dynamically
updated before they're loaded.