Apply hierarchical sort to buy & withdraw lists for consistent UX (#1196)
The changes made in
[#565](https://github.com/Meridian59/Meridian59/pull/565) have made
organization in rooms and guild halls a far smoother experience - this
PR applies the same hierarchical sort to the 'buy' and 'withdraw' boxes
so it can be enjoyed by those who frequent vault-men and NPC merchants
alike.

Code-wise the changes are very straight-forward; it applies the
algorithm introduced by @CardiovascularMike in #565 >
`GotObjectContents`.

This PR was tested on every NPC willing to make sales in Barloque, as
well as Obert with both the 'buy' and 'withdraw' commands.