Added new colors for normal and true lute
Addd two different lute colors to diffentiate the three lute types (normal/fine/true).  Fine lute uses the original graphics, while the normal lute uses the brighter version and the true lute uses the reddish version.

For compatiblity with other scripts, the original 'music.bgf" and 'luteov.bgf' files are unchanged and are still used in 'finelute.kod'.

* Added 'truelute.bbg' and 'truelute.bgf'. Edited 'truelute.kod' to use them.
* Added 'basiclute.bbg' and 'basiclute.bgf'.  Edited 'lute.kod' to use them.

Note: when this commit was made, the .bgf files had to be added to run/localclient/resource folder to show properly.