feature: Expanded functionality for the ice creature (#923)
This PR addresses #428 by implementing much, if not all, of the
remaining functionality for the existing ice creature. The work is based
on existing assets and code clues, with the goal of making the ice
creature game-ready. However, it does not include decisions about where
the creature will appear in-game—those changes, along with any
adjustments from play-testing, will be addressed in future updates.

To spawn a ice creature, use `dm monster ice creature`

### Key Updates
* Previously, the ice creature had a single variant carrying a sword.
This update utilizes unused assets to add functionality for a spear
weapon, which the unarmed variant can use for both melee and projectile
attacks. When throwing the projectile spear, there's a cooldown period
before another spear is generated.
* A new corpse asset has been created, as one was not present.
* Core elements such as description, difficulty, karma, treasure lists,
and AI have been updated to reflect a higher-karma, higher-difficulty
monster.

## Asset work
The original assets were extremely small, so they were extracted,
upscaled, and recompiled.

## Media

### Demo Video

https://drive.google.com/file/d/10eG3SJ0O9puEcaScAAKkhThEJk7G3pUJ/view?usp=sharing

### Unarmed

![image](https://github.com/user-attachments/assets/d387d689-ee88-470d-8cd8-cf3914547ab6)

### With Sword

![image](https://github.com/user-attachments/assets/8516aab0-5be4-42fc-a03c-5e686ebd8926)

### Walkaround Clip

https://github.com/user-attachments/assets/d6c291d8-227b-42f0-999c-1f7ac31589e2

### Sword Fight Clip

https://github.com/user-attachments/assets/e362b230-f49e-486f-8e3a-24817f94e5b7

### Unarmed/Spear Fight Clip

https://github.com/user-attachments/assets/b585b491-18e3-4de6-b58c-7547122a768a

Closes #428, Closes #427